analytics/requirements.txt
Django>=4.0
djangorestframework
kafka-python
